+++ connecteur/s2low/S2low.class.php
@@ -414,7 +414,12 @@ class S2low  extends TdtConnecteur {
  			array_shift($ligne);
 			$this->arActes = utf8_decode(implode("\n",$ligne));
 		}
-		
+
+		if ($result == -1){
+			array_shift($ligne);
+			array_shift($ligne);
+			$this->lastError = utf8_decode(implode("\n",$ligne));;
+		}
 		return $result;
 	}